spacelink-enterprise-gateway/api-gateway/requirements.txt
# SpaceLink Enterprise Gateway - API Dependencies

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2

# Database
sqlalchemy==2.0.35

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP Client
requests==2.32.3
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
